Fix #137, #114: Breaking change: Upgrade to Scala.js 0.6.18.
Also upgrade to sbt-web-scalajs 1.0.5, which has been updated for
Scala.js 0.6.17+.

The deprecated *launchers* of Scala.js are now completely
disregarded by scalajs-bundler. Projects that have a main method
must now use the Scala.js setting

  scalaJSUseMainModuleInitializer := true

so that the call to the `main` method is included in the
`-fastopt.js` and `-opt.js` files. Failure to do so will result in
a basically empty bundle that does not do anything.

On the other hand, projects that were exposed as a "library" with
several entry points, as explained in the cookbook, do not need to
override `scalaJSLauncher` anymore. Instead, they just have to
leave `scalaJSUseMainModuleInitializer` to its default value
`false`.