Implement webpack library bundling mode
This patch introduces a new configuration key, `webpackBundlingMode`,
which replaces the functionality previously provided by `enableReloadWorkflow`.
It builds upon what reload workflow was doing, in that it extracts the imported
dependencies from the scala.js output and generates a fake entrypoint file
which is used to run webpack against.

Then, instead of manually concatenating the outputs like reload workflow, it
instead simply builds the webpack module as a library and implements a trivial
require function that can be used to get at those dependencies.

This allows us to trigger webpack rebuilds indepndently of scala.js builds,
only when necessary, which cascades the same benefit down to the browser
and its ability to keep the libraries hot in cache while reloading the application
as it changes, and vice versa.