Introduce JS source directories.
A new key [jsSourceDirectories} is introduced. All files under thoses directories
will be bundled. If does not directly fixes #201, but offers a cleaner way to have
both js and non js files collected in the bundle. It also pave the way to remove
the dependency on `scalaJSNativeLibraries` that will go away in Scala.js 1.x