Upgrade to Scala.js 1.0.0-RC1 (and 0.6.31).
This includes the minimal changes required to upgrade the Scala.js
1.x support to 1.0.0-RC1, dropping 1.0.0-M7. The main limitation
is that we make sbt-scalajs-bundler directly depend on
scalajs-linker, instead of loading scalajs-linker by reflection
like sbt-scalajs does. This preserves the status quo of 1.0.0-M8
instead of benefitting from the improvements of 1.0.0-RC1 in that
area, so it could use improvements in the future but is correct for
the moment.

Since sbt-scalajs 1.0.0-RC1 requires at least sbt 1.2.1, we bump
sbt 1.0.2 to 1.2.1 in the build and in tests.

We upgrade to Scala.js 0.6.31 as well so that we can use ScalaTest
3.1.0 final.