With Scala.js 1.x, do not relink to compute the imported modules.
Scala.js 1.x offers a hook to customize the classpath of the linker
as well as how linkers are instantiated. We leverage this hook to
use a custom linker that computes the imported modules on the fly,
between the frontend and backend, and stores that in a file. We
then directly read this file instead of relinking from scratch.

We introduce a new private task `scalaJSBundlerImportedModules` to
abstract away the differences between 0.6.x and 1.x. It is computed
differently in the Scala.js-version-dependent `Settings.scala`.

The custom linker is in a new project, only used with Scala.js 1.x.

sbt-scalajs-bundler must still depend on scalajs-linker because it
uses the JavaScript trees defined in the latter.