Use `Path`s instead of Scala.js' virtual files in `NpmUpdateTasks`.
This basically consists in inlining the definitions from `compat`
for Scala.js 1.x at call site, even for 0.6.x. Since they only
depend on JDK `Path`s and the Java library `jimfs`, these also
work with Scala.js 0.6.x. In the process, we take the latest
developments of scala-js/jsdependencies into
`ScalaJSNativeLibraries`, to better accomodate the use of `Path`s.

This allows to remove the last remaining bits of `compat`.