Upgrade to sbt 1.3.8 and use Scala.js 1.x by default.
Upgrading to sbt 1.x required upgrading all sbt plugins.

Using Scala.js 1.x by default provides a better experience in
Metals out of the box, because of the `scalajs-bundler-linker`
project which cannot resolve with Scala.js 0.6.x.

The sbt plugin is still compiled against sbt 1.2.8, so that users
of sbt-scalajs-bundler are not forced to upgrade to sbt 1.3.x
themselves.