Load a newer version of the linker to match the version of sbt-scalajs.
With Scala.js 1.x, scalajs-bundler loads a custom linker with a
custom backend in `Global / scalaJSLinkerImpl / fullClasspath`.
Since the artifact with the custom linker depends on
`scalajs-linker`, we also receive that one. However, that always
loads the version of `scalajs-linker` against which scalajs-bundler
was built. If we combine it with a newer version of `sbt-scalajs`,
we still loaded the old version of `scalajs-linker`. This can
prevent scalajs-bundler from being used with non forward compatible
versions of Scala.js, or simply to miss out on bug fixes and
improvements from upstream.

This commit solves this issue by adding an explicit dependency on
scalajs-linker with the same version as that of sbt-scalajs (which
we find in `scalaJSVersion`). As long as that version of the linker
is backward binary compatible with our usage in the custom linker,
this will allow scalajs-bundler to be used with more recent
versions of sbt-scalajs.

We test this with a test that specifically bumps sbt-scalajs to
1.0.1 and contains a test for a bugfix introduced in the upstream
linker version 1.0.1.