Test only one configuration in AppVeyor.
The free tier of AppVeyor never tests configurations in parallel.
Since each configuration takes half an hour to test, running them
all sequentially takes a very long time.

Since it is extremely unlikely that only one of the configurations
would fail, only on Windows, we rely on the Linux builds to test
the other configurations, and only test one of them on Windows.