Upgrade to Scala.js 1.3.0.
This commit contains the minimum amount of changes for
scalajs-bundler to work with Scala.js 1.3.0. However, it still only
works when module splitting is not actually used. To be precise, it
still relies on the legacy keys `fastOptJS` and `fullOptJS`, and
therefore on the fact that linking only produces one module.

Integrating with module splitting is left for future work.

We add a main method in the yarn-interactive test, because
otherwise it creates an empty Scala.js `Report`, and apparently the
`fastOptJS`/`fullOptJS` compatibility setup in Scala.js 1.3.0 does
not like that.