Repair the `manual`, reintroducing referenced scripted tests.
The scripted tests `global-namespace-with-jsdom-unit-testing_sjs-0.6`
and `webpack-assets-cookbook_sjs-0.6` were deleted in
08d6e4802bd8826fdca29ec659509e7d079763aa because they were only
working in Scala.js 0.6.x. However, they were referenced by the
manual, in `cookbook.md`, which broken the manual build. We did not
notice because the CI was not configured to build the manual on
every pull request.

We reintroduce those tests, using updated dependencies that support
Scala.js 1.x (for `webpack-assets-cookbook`) or inline facades
emulating the old dependencies (for
`global-namespace-with-jsdom-unit-testing`).

This allows `manual/ornate` to successfully run again. We add it to
the CI to prevent future regressions.